+++ tests/test_sparkpost_integration.py
@@ -7,13 +7,12 @@ from django.test import SimpleTestCase, override_settings, tag
 from anymail.exceptions import AnymailAPIError
 from anymail.message import AnymailMessage
 
-from .utils import AnymailTestMixin, sample_image_path, RUN_LIVE_TESTS
+from .utils import AnymailTestMixin, sample_image_path
 
 SPARKPOST_TEST_API_KEY = os.getenv('SPARKPOST_TEST_API_KEY')
 
 
 @tag('sparkpost', 'live')
-@unittest.skipUnless(RUN_LIVE_TESTS, "RUN_LIVE_TESTS disabled in this environment")
 @unittest.skipUnless(SPARKPOST_TEST_API_KEY,
                      "Set SPARKPOST_TEST_API_KEY environment variable "
                      "to run SparkPost integration tests")
